requirements.txt
asgiref==3.4.1
astroid==2.8.3
Django==3.2.8
django-filter==21.1
django-plaintext-password==0.1.0
djangorestframework==3.12.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
psycopg2-binary==2.9.1
pylint==2.11.1
pylint-plugin-utils==0.6
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.13.2
